refactor(playtomic): split 2236-LOC page into focused module
Extract app/rdb/playtomic/page.tsx into components/playtomic/ following
the same convention as app-shell/, health/, travel/, rh/, and tasks/.

Behavior preserved 1:1. Page is now a thin RequireAccess wrapper that
delegates to PlaytomicView.
